+++ cogs/admin/admin.py
@@ -134,6 +134,29 @@ class Administration(commands.Cog):
                     print(f'[{Time.timeCST()}] [Roundtrip: {Roundtrip.rt(self)}ms.] CONSOLE | LOG - UNBAN: A user was unbanned in #{ctx.channel}! [{member}]')
                     return  
 
+    # MUTE METHOD
+    @commands.command(help = "Adds a mute role to a person for a specified duration or none.")
+    @commands.has_any_role('Admin','House of Lords')
+    async def mute(self, ctx, member: discord.Member, duration: DurationConverter, *, reason = None):
+        Mute = '843662762600431637'
+        if (member == ctx.author):
+            await ctx.reply("Listen, can you not mute yourself...?")
+        else:
+            multiplier = {'s': 1, 'm': 60, 'h': 3600, 'd': 86400}
+            amount, unit = duration
+
+            if (reason): rb = f', with the reason being {str.capitalize(reason)}'
+            else: rb = '.'
+
+            em = discord.Embed (
+                description = f'**{member}** has been muted for {amount}{unit}{rb}'
+            )
+            em.set_footer (
+                text = f'Tachibana Administration Protocol: {Time.dateTimeUTC()}'
+            )
+            await member.add_roles(roles = Mute, reason = reason)
+            await asyncio.sleep(amount * multiplier[unit])
+            await member.remove_roles(roles = Mute, reason = 'The specified time is over.')
     # # ADD ROLES - METHOD
     # @commands.command(help = 'Adds a role to a user.')
     # @commands.has_permissions(ban_members = True)